scouting builder moved to builder
