+++ zenoh/src/api/builders/mod.rs
@@ -19,5 +19,6 @@ pub(crate) mod query;
 pub(crate) mod queryable;
 pub(crate) mod reply;
 pub(crate) mod sample;
+pub(crate) mod scouting;
 pub(crate) mod session;
 pub(crate) mod subscriber;
